image information easily changed, adds docker pull
This allows changes in the openshift image to be used to be passed from
the Vagrantfile into the box.

docker pull is now done in sccli to ensure images for requested version
are present.

Uses IMAGE_* variables so that sccli can use similar code for future
services.  Also this represents that not all versions of OpenShift are
named Origin.

DOCKER_REGISTRY was removed from openshift_options and registry was made
part of the image name

`docker tag` removed from the unit file as it was only needed in one
place

When starting openshift, enable the service for future reboots.

CLI binaries now copied with `docker cp` for efficiency

There is a second PR that makes changes to the Vagrant file in the ADB